setup MainViewController
