+++ Navigation/Models/Post.swift
@@ -8,6 +8,5 @@
 import Foundation
 
 struct Post {
-    
     var title: String
 }
