+++ Navigation/NavigationControllers/InfoViewController.swift
@@ -11,11 +11,11 @@ class InfoViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        makeButtin()
+        makeButton()
         view.backgroundColor = .red
     }
     
-    private func makeButtin() {
+    private func makeButton() {
         let button = UIButton(frame: CGRect(x: 0, y: 0, width: 200, height: 40))
         button.center = view.center
         button.setTitle("Back", for: .normal)
